LG-7187: Send survey when user shares info with partner; add fallback test URL for survey
